+++ CHANGELOG.md
@@ -5,12 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.4] - 2025-01-30
 
 ### Added
 
 - Action button allows user to export tabular results to file.
 - Tool call allows user to ask the model to export tabular results to file.
+
+### Changed
+
+- Refreshing browser or clicking "Clear chat" button erases data export 
+table and re-initialises chat interface.
 
 ## [0.2.3] - 2025-01-24